+++ src/chains.js
@@ -8,13 +8,33 @@ const { getChainDefaults } = require('./config');
 
 const debug = Debug('iexec:chains');
 
+const CHAIN_ALIASES_MAP = {
+  1: 'mainnet',
+  5: 'goerli',
+  133: 'viviani',
+  134: 'bellecour',
+};
+
+const CHAIN_NAME_MAP = {
+  1: { id: 1, flavour: 'standard' },
+  mainnet: { id: 1, flavour: 'standard' },
+  5: { id: 5, flavour: 'standard' },
+  goerli: { id: 5, flavour: 'standard' },
+  133: { id: 133, flavour: 'standard' },
+  viviani: { id: 133, flavour: 'standard' },
+  134: { id: 134, flavour: 'standard' },
+  bellecour: { id: 134, flavour: 'standard' },
+  enterprise: { id: 1, flavour: 'enterprise' },
+  'enterprise-sidechain': { id: 133, flavour: 'enterprise' },
+};
+
 const createChainFromConf = (
   chainName,
   chainConf,
   { bridgeConf, providersOptions } = {},
 ) => {
   try {
-    const chain = Object.assign({}, chainConf);
+    const chain = { ...chainConf };
     const provider = getDefaultProvider(chainConf.host, providersOptions);
     chain.name = chainName;
     const contracts = new IExecContractsClient({
@@ -22,6 +42,7 @@ const createChainFromConf = (
       chainId: chain.id,
       hubAddress: chain.hub,
       isNative: chain.native,
+      flavour: chain.flavour,
     });
     chain.contracts = contracts;
     if (bridgeConf) {
@@ -34,10 +55,8 @@ const createChainFromConf = (
         provider: bridgeProvider,
         chainId: bridgeConf.id,
         hubAddress: bridgeConf.hub,
-        isNative:
-          bridgeConf.native === undefined
-            ? !chain.contracts.isNative
-            : bridgeConf.native,
+        isNative: bridgeConf.native,
+        flavour: bridgeConf.flavour,
       });
     }
     return chain;
@@ -59,14 +78,11 @@ const loadChain = async (chainName, { spinner = Spinner() } = {}) => {
         loadedConf = chainsConf.chains[chainName];
         name = chainName;
       } else {
-        const names = Object.keys(chainsConf.chains);
-        names.forEach((n) => {
-          const chainConf = chainsConf.chains[n];
-          if (chainConf.id && chainConf.id === chainName) {
-            name = n;
-            loadedConf = chainConf;
-          }
-        });
+        const alias = CHAIN_ALIASES_MAP[chainName];
+        if (alias && chainsConf.chains[alias]) {
+          loadedConf = chainsConf.chains[alias];
+          name = alias;
+        }
         if (!name) throw Error(`Missing "${chainName}" chain in "chain.json"`);
       }
     } else if (chainsConf.default) {
@@ -82,11 +98,18 @@ const loadChain = async (chainName, { spinner = Spinner() } = {}) => {
     }
     if (!name) throw Error('Missing chain parameter. Check your "chain.json" file');
 
-    const defaultConf = getChainDefaults(loadedConf.id);
+    const idAndFlavour = {
+      ...CHAIN_NAME_MAP[name],
+      ...(loadedConf.id && { id: loadedConf.id }),
+      ...(loadedConf.flavour && { flavour: loadedConf.flavour }),
+    };
+
+    const defaultConf = getChainDefaults(idAndFlavour);
+
     debug('loading chain', name);
     debug('loadedConf', loadedConf);
     debug('defaultConf', defaultConf);
-    const conf = { ...defaultConf, ...loadedConf };
+    const conf = { ...idAndFlavour, ...defaultConf, ...loadedConf };
     debug('conf', conf);
     if (!conf.host) {
       throw Error(
@@ -95,25 +118,36 @@ const loadChain = async (chainName, { spinner = Spinner() } = {}) => {
     }
 
     let bridgeConf;
-    if (conf.bridge && conf.bridge.bridgedChainId) {
-      const names = Object.keys(chainsConf.chains);
-      names.forEach((n) => {
-        const bridgeLoadedConf = chainsConf.chains[n];
-        if (
-          bridgeLoadedConf.id
-          && bridgeLoadedConf.id === conf.bridge.bridgedChainId
-        ) {
-          const bridgeDefaultConf = getChainDefaults(bridgeLoadedConf.id);
-          debug('bridgeLoadedConf', bridgeLoadedConf);
-          debug('bridgeDefaultConf', defaultConf);
-          bridgeConf = { ...bridgeDefaultConf, ...bridgeLoadedConf };
-          if (!bridgeConf.host) {
-            throw Error(
-              `Missing RPC host for bridged chain, no "host" key in "chain.json" and no default value for bridged chain ${bridgeConf.id}`,
-            );
-          }
+    const bridgedChainNameOrId = conf.bridge.bridgedChainName || conf.bridge.bridgedChainId;
+    if (conf.bridge && bridgedChainNameOrId) {
+      let bridgeLoadedConf;
+      if (chainsConf.chains[bridgedChainNameOrId]) {
+        bridgeLoadedConf = chainsConf.chains[bridgedChainNameOrId];
+      } else {
+        const alias = CHAIN_ALIASES_MAP[bridgedChainNameOrId];
+        if (alias && chainsConf.chains[alias]) {
+          bridgeLoadedConf = chainsConf.chains[alias];
         }
-      });
+        if (!bridgeLoadedConf) throw Error(`Missing "${chainName}" chain in "chain.json"`);
+      }
+      const bridgedIdAndFlavour = {
+        ...CHAIN_NAME_MAP[bridgedChainNameOrId],
+        ...(bridgeLoadedConf.id && { id: bridgeLoadedConf.id }),
+        ...(bridgeLoadedConf.flavour && { flavour: bridgeLoadedConf.flavour }),
+      };
+      const bridgeDefaultConf = getChainDefaults(bridgedIdAndFlavour);
+      debug('bridgeLoadedConf', bridgeLoadedConf);
+      debug('bridgeDefaultConf', defaultConf);
+      bridgeConf = {
+        ...bridgedIdAndFlavour,
+        ...bridgeDefaultConf,
+        ...bridgeLoadedConf,
+      };
+      if (!bridgeConf.host) {
+        throw Error(
+          `Missing RPC host for bridged chain, no "host" key in "chain.json" and no default value for bridged chain ${bridgeConf.id}`,
+        );
+      }
     }
     debug('bridged chain', bridgeConf);
     const chain = createChainFromConf(name, conf, {